requirements.txt
openai
hydra-core
loguru
streamlit
matplotlib
pandas
numpy
tensorboard
Pillow 
opencv-python-headless
torchviz
graphviz